# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM API для управления заявками бизнеса через Telegram, Instagram и MAX
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export type LeadPlatform = (typeof LeadPlatform)[keyof typeof LeadPlatform];

export const LeadPlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export type LeadStatus = (typeof LeadStatus)[keyof typeof LeadStatus];

export const LeadStatus = {
  hot: "hot",
  warm: "warm",
  cold: "cold",
} as const;

export interface Lead {
  id: number;
  clientName: string;
  platform: LeadPlatform;
  service: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  comment?: string | null;
  status: LeadStatus;
  /** @nullable */
  recommendation?: string | null;
  isPriority: boolean;
  createdAt: string;
  updatedAt: string;
}

export type CreateLeadBodyPlatform =
  (typeof CreateLeadBodyPlatform)[keyof typeof CreateLeadBodyPlatform];

export const CreateLeadBodyPlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export type CreateLeadBodyStatus =
  (typeof CreateLeadBodyStatus)[keyof typeof CreateLeadBodyStatus];

export const CreateLeadBodyStatus = {
  hot: "hot",
  warm: "warm",
  cold: "cold",
} as const;

export interface CreateLeadBody {
  clientName: string;
  platform: CreateLeadBodyPlatform;
  service: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  comment?: string | null;
  status: CreateLeadBodyStatus;
  /** @nullable */
  recommendation?: string | null;
  isPriority?: boolean;
}

export type UpdateLeadBodyPlatform =
  (typeof UpdateLeadBodyPlatform)[keyof typeof UpdateLeadBodyPlatform];

export const UpdateLeadBodyPlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export type UpdateLeadBodyStatus =
  (typeof UpdateLeadBodyStatus)[keyof typeof UpdateLeadBodyStatus];

export const UpdateLeadBodyStatus = {
  hot: "hot",
  warm: "warm",
  cold: "cold",
} as const;

export interface UpdateLeadBody {
  clientName?: string;
  platform?: UpdateLeadBodyPlatform;
  service?: string;
  /** @nullable */
  details?: string | null;
  /** @nullable */
  quantity?: string | null;
  /** @nullable */
  deadline?: string | null;
  /** @nullable */
  price?: string | null;
  /** @nullable */
  comment?: string | null;
  status?: UpdateLeadBodyStatus;
  /** @nullable */
  recommendation?: string | null;
  isPriority?: boolean;
}

export type UpdateLeadStatusBodyStatus =
  (typeof UpdateLeadStatusBodyStatus)[keyof typeof UpdateLeadStatusBodyStatus];

export const UpdateLeadStatusBodyStatus = {
  hot: "hot",
  warm: "warm",
  cold: "cold",
} as const;

export interface UpdateLeadStatusBody {
  status: UpdateLeadStatusBodyStatus;
}

export interface AnalyticsSummary {
  totalLeads: number;
  hotLeads: number;
  warmLeads: number;
  coldLeads: number;
  /** @nullable */
  averagePrice?: number | null;
  priorityLeads: number;
  todayLeads: number;
}

export interface PlatformCount {
  platform: string;
  count: number;
}

export interface StatusCount {
  status: string;
  count: number;
}

export interface AnalyzeTariffBody {
  /** Произвольный текст с описанием товаров/услуг и ценами */
  rawText: string;
  /**
   * Тип бизнеса (например, кондитерская, фотограф)
   * @nullable
   */
  businessType?: string | null;
  /** JSON array of selected platforms e.g. '["Telegram","Instagram","MAX"]' */
  platforms?: string;
}

export interface TariffItem {
  name: string;
  /** @nullable */
  description?: string | null;
  price: string;
  /** @nullable */
  unit?: string | null;
  /** @nullable */
  category?: string | null;
}

export interface AnalyzeTariffResponse {
  businessType: string;
  items: TariffItem[];
  summary: string;
  /** Готовый промпт для бота, который будет общаться с клиентами */
  botPrompt: string;
}

export interface TariffSettings {
  id: number;
  rawText: string;
  businessType: string;
  structuredData: string;
  botPrompt: string;
  /** JSON array of selected platforms e.g. '["Telegram","Instagram","MAX"]' */
  platforms: string;
  createdAt: string;
  updatedAt: string;
}

export interface SaveTariffSettingsBody {
  rawText: string;
  businessType: string;
  structuredData: string;
  botPrompt: string;
  /** JSON array of selected platforms e.g. '["Telegram","Instagram","MAX"]' */
  platforms: string;
}

export type BotAccountPlatform =
  (typeof BotAccountPlatform)[keyof typeof BotAccountPlatform];

export const BotAccountPlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export interface BotAccount {
  id: number;
  platform: BotAccountPlatform;
  accountName: string;
  accountHandle: string;
  notificationChat: string;
  isActive: boolean;
  createdAt: string;
  updatedAt: string;
}

export type SaveBotAccountBodyPlatform =
  (typeof SaveBotAccountBodyPlatform)[keyof typeof SaveBotAccountBodyPlatform];

export const SaveBotAccountBodyPlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export interface SaveBotAccountBody {
  platform: SaveBotAccountBodyPlatform;
  accountName: string;
  accountHandle: string;
  notificationChat: string;
  isActive?: boolean;
}

export type LeadChatMessagePlatform =
  (typeof LeadChatMessagePlatform)[keyof typeof LeadChatMessagePlatform];

export const LeadChatMessagePlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export interface LeadChatMessage {
  id: number;
  /** @nullable */
  leadId?: number | null;
  platform: LeadChatMessagePlatform;
  title: string;
  message: string;
  createdAt: string;
}

export interface RegisterUserBody {
  /** Telegram username без @ (например misha_shop) */
  telegramUsername: string;
}

export interface UserProfile {
  id: number;
  telegramUsername: string;
  apiToken: string;
  /** @nullable */
  telegramChatId?: string | null;
  createdAt: string;
  updatedAt: string;
}

export type ListLeadsParams = {
  status?: ListLeadsStatus;
  platform?: ListLeadsPlatform;
  dateFrom?: string;
  dateTo?: string;
};

export type ListLeadsStatus =
  (typeof ListLeadsStatus)[keyof typeof ListLeadsStatus];

export const ListLeadsStatus = {
  hot: "hot",
  warm: "warm",
  cold: "cold",
} as const;

export type ListLeadsPlatform =
  (typeof ListLeadsPlatform)[keyof typeof ListLeadsPlatform];

export const ListLeadsPlatform = {
  Telegram: "Telegram",
  Instagram: "Instagram",
  MAX: "MAX",
} as const;

export type GetUserProfileParams = {
  apiToken: string;
};