# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: CRM API для управления заявками бизнеса через Telegram, Instagram и MAX
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users/register:
    post:
      operationId: registerUser
      tags: [users]
      summary: Зарегистрировать пользователя и получить API-токен
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterUserBody"
      responses:
        "200":
          description: Пользователь зарегистрирован
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "400":
          description: Ошибка валидации
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /users/profile:
    get:
      operationId: getUserProfile
      tags: [users]
      summary: Получить профиль текущего пользователя по токену
      parameters:
        - name: apiToken
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Профиль пользователя
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "404":
          description: Пользователь не найден
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    RegisterUserBody:
      type: object
      properties:
        telegramUsername:
          type: string
          description: Telegram username без @ (например misha_shop)
      required:
        - telegramUsername
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        telegramUsername:
          type: string
        apiToken:
          type: string
        telegramChatId:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - telegramUsername
        - apiToken
        - createdAt
        - updatedAt