# lib/api-zod/src/generated/types/registerUserBody.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM API для управления заявками бизнеса через Telegram, Instagram и MAX
 * OpenAPI spec version: 0.1.0
 */

export interface RegisterUserBody {
  /** Telegram username без @ (например misha_shop) */
  telegramUsername: string;
}