# lib/api-zod/src/generated/types/userProfile.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * CRM API для управления заявками бизнеса через Telegram, Instagram и MAX
 * OpenAPI spec version: 0.1.0
 */

export interface UserProfile {
  id: number;
  telegramUsername: string;
  apiToken: string;
  /** @nullable */
  telegramChatId?: string | null;
  createdAt: string;
  updatedAt: string;
}